// src/enums/cacheEnums.ts
export enum CacheEnums {
  // 本地缓冲key
  // access_token
  ACCESS_TOKEN_KEY = 'access_token',
  // refresh_token
  REFRESH_TOKEN_KEY = 'refresh_token',
  // access_token过期时间
  ACCESS_TOKEN_EXPIRES_IN_KEY = 'access_token_expires_in',
  // 获取access_token时间
  ACQUIRE_TOKEN_TIME_KEY = 'acquire_token_time',
  // 账号
  ACCOUNT_KEY = 'account',
  // 设置
  SETTING_KEY = 'setting'
}